refactor: rewrite markdownConverter to ts
